connection pool admin
git-svn-id: svn://svn.caucho.com/resin/trunk@6026 9c94448d-38f1-0310-a231-d98308ff1ebf

+++ modules/resin/src/com/caucho/admin/AverageTimeProbe.java
@@ -54,13 +54,13 @@ public final class AverageTimeProbe extends Probe {
   /**
    * Return the probe's next sample.
    */
-  public final long sample()
+  public final double sample()
   {
     long sum = _sum.getAndSet(0);
     int count = _count.getAndSet(0);
 
     if (count != 0)
-      return sum / count;
+      return sum / (double) count;
     else
       return 0;
   }
